Change scrollbar behaviour and change theme
The theme is set to be dark with white accents so this commit realizes
this. I changed the background to something grey and the fonts to white
as well as the links after they have been clicked. I also modified the
scrollbar behaviour to not show the scrollbar as it distracts from the
name display.